requirements.txt
bcrypt==4.1.2
blinker==1.8.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.2
paramiko==3.4.0
psycopg2==2.9.9
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sshtunnel==0.4.0
tzdata==2024.1
uuid==1.30
Werkzeug==3.0.2
